refactor: move uploadUserAvatar endpoint to AccountController
